docs: updated docs for mentions plugin and added examples for mentions
