+++ development/text-editor-app/components/text-editor/EnabledMentions/EnabledMentions.tsx
@@ -1,12 +1,45 @@
 import React, { useState } from "react";
+import { LexicalEditor } from "lexical";
 import { TextEditor } from "@locoworks/reusejs-react-text-editor";
 import "@locoworks/reusejs-react-text-editor/css";
 
-const Example = () => {
-	const [editable, setEditable] = useState(false);
-	const [data, setData] = useState("Here");
+const EnabledMentions = () => {
+	const [editable, setEditable] = useState<boolean>(false);
+	const [data, setData] = useState<string | TrustedHTML>("Here");
 
-	function convertFilesToImageUrl(files) {
+	const dummyMentionsData = [
+		"Aayla Secura",
+		"Adi Gallia",
+		"Niima the Hutt",
+		"Nines",
+		"Norra Wexley",
+		"Nute Gunray",
+		"Val Beckett",
+		"Vanden Willard",
+		"Vice Admiral Amilyn Holdo",
+		"Vober Dand",
+		"WAC-47",
+		"Wedge Antilles",
+		"Wes Janson",
+		"Wicket W. Warrick",
+		"Wilhuff Tarkin",
+		"Wollivan",
+		"Wuher",
+		"Wullf Yularen",
+		"Xamuel Lennox",
+		"Yaddle",
+		"Yarael Poof",
+		"Yoda",
+		"Zam Wesell",
+		"Zev Senesca",
+		"Ziro the Hutt",
+		"Zuckuss",
+	].map((name) => ({
+		mentionName: `user_${name}`,
+		label: name,
+	}));
+
+	function convertFilesToImageUrl(files: FileList | null) {
 		if (!files || files.length === 0) {
 			return null;
 		}
@@ -21,7 +54,7 @@ const Example = () => {
 		return imageUrls.length > 0 ? imageUrls : null;
 	}
 
-	function onChange(_editorRef, payload) {
+	function onChange(_editorRef: LexicalEditor | null, payload: any) {
 		setData(payload["html"]);
 	}
 
@@ -30,6 +63,7 @@ const Example = () => {
 			<TextEditor
 				editable={editable}
 				setEditable={setEditable}
+				mentionsData={dummyMentionsData}
 				convertFilesToImageUrl={convertFilesToImageUrl}
 				onChangeCallback={onChange}
 			/>
@@ -46,4 +80,4 @@ const Example = () => {
 	);
 };
 
-export default Example;
+export default EnabledMentions;
