+++ development/text-editor-app/pages/text-editor/index.mdx
@@ -5,6 +5,8 @@ author: "Upneet Kaur Hunjan"
 ---
 import Example from "../../components/text-editor/Example/Example.tsx";
 import PrepopulatedTextEditor from "../../components/text-editor/PrepopulatedTextEditor/PrepopulatedTextEditor.tsx";
+import EnabledMentions from "../../components/text-editor/EnabledMentions/EnabledMentions.tsx";
+import EnabledMentionsWithCustomLookup from "../../components/text-editor/EnabledMentionsWithCustomLookup/EnabledMentionsWithCustomLookup.tsx";
 
 ## Text-Editor
 
@@ -43,8 +45,7 @@ import { TextEditor } from "@locoworks/reusejs-react-text-editor";
 ## Types/ Exported Components
 
 The `TextEditor` component accepts the following props:
-
-- `mentionsData`: An array of objects representing mentions data.
+- `mentionsData`: An optional array of objects representing mentions data. It indicates enabling or disabling the mentions.Its value is "false" by default which means disabled mentions.
 - `useMentionLookupService`: An optional function to handle mention lookup service and return results for search.
 - `convertFilesToImageUrl`: A function to convert files to image URLs and return Array of imageUrls.
 - `onChangeCallback`: An optional callback for editor reference changes and payload.
@@ -66,6 +67,16 @@ The `TextEditor` component accepts the following props:
 ```js path=../../components/text-editor/PrepopulatedTextEditor name=PrepopulatedTextEditor
 ```
 
+### Enabled Mentions 
+<EnabledMentions/>
+```js path=../../components/text-editor/EnabledMentions name=EnabledMentions
+```
+
+### Enabled Mentions with Custom lookup function
+<EnabledMentionsWithCustomLookup/>
+```js path=../../components/text-editor/EnabledMentionsWithCustomLookup name=EnabledMentionsWithCustomLookup
+```
+
 ## Styling
 
 The styling of the editorWrapper can be customized using Tailwind CSS .
@@ -84,12 +95,12 @@ import "@locoworks/reusejs-react-text-editor/css";
 
 | Prop                  | Type                                 | Description                                                                                                       |
 | --------------------- | ------------------------------------ | ----------------------------------------------------------------------------------------------------------------- |
-| `mentionsData	` | `Array<{ mentionName: string; label: string; }>`                             | An array of objects representing mentions data. Each object should have mentionName and label properties.                                                           |
+| `mentionsData	` | `Array<{ mentionName: string; label: string; }> \| false`                             | An optional array of objects representing mentions data. Each object should have mentionName and label properties. Its false by default.                                                           |
 | `useMentionLookupService` | `(mentionString: string \| null, mentionsData) => Array<{ mentionName: string; label: string; }>`                             | An optional function to handle mention lookup and return results.                                                           |
 | `convertFilesToImageUrl`   | `(files: FileList \| null) => Array<string> \| null`                             | A function to convert files to image URLs and return image URLs.                                                  |
 | `onChangeCallback`       | `(editorRef: LexicalEditor \| null, payload: any) => void`                             | An optional callback for editor reference changes and payload.                                                    |
-| `wrapperClass`           | `String`                               | An optional class for styling the editor wrapper.                                                                  |
-| `editable`               | `Boolean`                              | A boolean indicating the edit state of the editor.                                                                |
+| `wrapperClass`           | `string`                               | An optional class for styling the editor wrapper.                                                                  |
+| `editable`               | `boolean`                              | A boolean indicating the edit state of the editor.                                                                |
 | `setEditable`            | `React.Dispatch<React.SetStateAction<boolean>>`                             | An optional function to set the editable state of the editor.                        |
 | `htmlData`            | `string`                             | An optional string to set the initial state of the editor.                                                       |
 | `placeholderText`            | `string`                             | An optional string for the placeholder inside the editor. Its value is "Start Typing..." by default                                                       |
@@ -97,6 +108,8 @@ import "@locoworks/reusejs-react-text-editor/css";
 ## Notes
 >The TextEditor component uses the Lexical library for rich text editing.
 Check the Lexical Composer documentation for advanced configuration options.
+
+>To enable the mentions plugin , you have to pass the mentionsData which is false by default and thus disabled.
 
 export default ({ children }) => {
   return <div className="prose max-w-full mx-4 pb-10" >{children}</div>
